requirements.txt
polars
openai
reportlab
python-dotenv
matplotlib
seaborn
pandas
